+++ scripts/hmm_general_lib.py
